Fixed inventory dropping bug
Fixed a bug where clicking on the drop/trash button multiple times while waiting on a amount would sometimes result in a crash. Now, it simply acts as if the amount entered was 1. Also preventing the player from opening both the trash and drop options at the same time (as that would likely also cause an error)